+++ ms-user/app/UserController.py
@@ -1,4 +1,5 @@
-from app.models.User import UserOut, UserUpdate
+from app.models.UserUpdate import UserUpdate
+from app.models.UserOut import UserOut
 from app.UserService import UserService
 from fastapi import FastAPI, Request, UploadFile, status
 from typing import List
@@ -9,8 +10,7 @@ us = UserService()
 
 @app.get("/{user_id}", status_code=status.HTTP_200_OK, response_model=UserOut)
 async def get_user(user_id: str) -> UserOut:
-    result = await us.get_user_by_id(user_id)
-    return UserOut(**result)
+    return await us.get_user_by_id(user_id)
 
 
 @app.get("/", status_code=status.HTTP_200_OK, response_model=UserOut)
@@ -21,8 +21,7 @@ async def search_user(username: str) -> UserOut:
 @app.patch("/", status_code=status.HTTP_200_OK, response_model=UserOut)
 async def update_user(request: Request, updated_user: UserUpdate) -> UserOut:
     user_id = us.extract_user_id_from_token(request)
-    result = await us.update_user_by_id(user_id, updated_user)
-    return UserOut(**result)
+    return await us.update_user_by_id(user_id, updated_user)
 
 
 @app.patch("/image", status_code=status.HTTP_200_OK, response_model=UserOut)
@@ -48,7 +47,7 @@ async def get_user_followers(user_id: str) -> List:
 
 
 @app.patch("/following/{user_id}", status_code=status.HTTP_200_OK, response_model=UserOut)
-async def follow_user(request: Request, user_id: str):
+async def follow_user(request: Request, user_id: str) -> UserOut:
     curr_user_id = us.extract_user_id_from_token(request)
     return await us.follow_user_by_id(curr_user_id, user_id)
